Make interface helpers public
This allows making the tests work on the public interface too.
